+++ src/macros/macrolib.js
@@ -11,7 +11,6 @@ import './macros/macrolib/actions';
 import './macros/macrolib/addclass-toggleclass';
 import './macros/macrolib/append-prepend-replace';
 import './macros/macrolib/audio-api-macros';
-import './macros/macrolib/back-return';
 import './macros/macrolib/button-link';
 import './macros/macrolib/capture';
 import './macros/macrolib/checkbox';